[Kodi] Add v16 folder; make v18 an alias for v17
 - at least for now v17 and v18 are equivalent
 - v16 is different and uses the "old" syntax
 - https://kodi.wiki/view/NFO_files/Movies is up-to-date for v17
