+++ MediaElch.pro
@@ -153,6 +153,7 @@ SOURCES += src/main.cpp \
     src/media_centers/kodi/ConcertXmlReader.cpp \
     src/media_centers/kodi/EpisodeXmlReader.cpp \
     src/media_centers/kodi/MovieXmlWriter.cpp \
+    src/media_centers/kodi/v16/MovieXmlWriterV16.cpp \
     src/media_centers/kodi/v17/MovieXmlWriterV17.cpp \
     src/media_centers/kodi/v18/MovieXmlWriterV18.cpp \
     src/media_centers/kodi/MovieXmlReader.cpp \
@@ -368,6 +369,7 @@ HEADERS  += Version.h \
     src/media_centers/kodi/ConcertXmlReader.h \
     src/media_centers/kodi/EpisodeXmlReader.h \
     src/media_centers/kodi/MovieXmlWriter.h \
+    src/media_centers/kodi/v16/MovieXmlWriterV16.h \
     src/media_centers/kodi/v17/MovieXmlWriterV17.h \
     src/media_centers/kodi/v18/MovieXmlWriterV18.h \
     src/media_centers/kodi/MovieXmlReader.h \
